Attempts to fix some deprecations
In PHP the switch statement is considered a looping structure for the purposes of continue. continue behaves like break (when no arguments are passed) but will raise a warning as this is likely to be a mistake. If a switch is inside a loop, continue 2 will continue with the next iteration of the outer loop.

Risks
--

1. Not much actual benefit in changing it, except it silences some warnings (wood/trees)
2. I couldn't figure out what's best to do with this one:
    PHP Warning:  "continue" targeting switch is equivalent to "break". Did you mean to use "continue 2"? in /Users/gavin/Code/bindhq-webapp/vendor/dompdf/dompdf/src/Css/Stylesheet.php on line 1360